move test-utils bits over to the test-utils package
