+++ packages/opentelemetry-test-utils/package.json
@@ -49,9 +49,10 @@
     "@opentelemetry/core": "^1.0.0",
     "@opentelemetry/exporter-jaeger": "^1.3.1",
     "@opentelemetry/instrumentation": "^0.44.0",
-    "@opentelemetry/sdk-trace-node": "^1.8.0",
     "@opentelemetry/resources": "^1.8.0",
+    "@opentelemetry/sdk-node": "^0.44.0",
     "@opentelemetry/sdk-trace-base": "^1.8.0",
+    "@opentelemetry/sdk-trace-node": "^1.8.0",
     "@opentelemetry/semantic-conventions": "^1.0.0"
   }
 }
